+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     in mode test disable html rendering and return template vars on controller call.
   * PFM_ADMIN_EMAIL: default super admin contact email
   * PFM_ADMIN_PASSWORD: default super admin user password (install only)
+  * PFM_AMQP_HOST, PFM_AMQP_PORT, PFM_AMQP_USER, PFM_AMQP_PASSWORD: rabbitmq settings (optional)
 
 * New config parameters:
   * sql_debug: combined with DEBUG mode, trace sql queries